requirements.txt
streamlit
pdfminer.six
scikit-learn
nltk